Issue with rekall binaries and symlinks
SIFT cli wants to install rekall as well, and fails when the binaries are present because it can't create a symlink. Since both the remnux and sift installs of rekall are the same at this time, there is no harm in removing rekall from remnux, removing the /usr/local/bin/rek* binaries and allowing the sift-cli to install.